Refactor run server internals
